requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
Django==1.9.6
django-celery==3.2.1
django-formtools==1.0
django-reversion==2.0.7
HTMLParser==0.0.2
httplib2==0.9.2
kombu==3.0.37
MarkupSafe==0.23
PyMySQL==0.7.9
pytz==2016.7
requests==2.11.1
SQLAlchemy==1.1.3
vine==1.1.3
